+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [0.0.4]
+
+-   Append target line of code to hover
+-   Fix `store` description
+
 ## [0.0.3]
 
 -   Added unconditional keyword (first 4 chars) matching